+++ examples/shared/src/main/scala-2/examples/SmartTypes.scala
@@ -3,8 +3,6 @@ package examples
 
 import zio.prelude._
 
-import java.util.UUID
-
 object SmartTypes extends App {
   import Assertion._
 
@@ -66,12 +64,6 @@ object SmartTypes extends App {
   }
   type Email = Email.Type
   val email: Email = Email("test@test.com")
-
-  object GenericItemId extends Newtype[UUID]
-  type GenericItemId = GenericItemId.Type
-
-  object SpecificItemId extends Subtype[GenericItemId]
-  type SpecificItemId = SpecificItemId.Type
 
 }
 // scalafix:on
